Proyecto Final Ecommers

+++ main.js
@@ -1,10 +1,4 @@
-const productos = [
-    {id: 1, burger: "Fresh Burger", price: 1030 , stock: true},
-    {id: 2,burger: "Special Bacon", price: 1330 , stock: false},
-    {id: 3,burger: "Le Blue", price: 1110 , stock: true},
-    {id: 4,burger: "Burger Stacker Quintuple", price: 1480 , stock: true},
-    {id: 5,burger: "Doble Cheese Burger", price: 1170 , stock: true}
-];
+
 
 let productosMayorPrecio = [];
 let productosMenorPrecio = [];
@@ -13,32 +7,36 @@ let productosEnElCarrito =[];
 let acumulador = ``;
 let padreCarrito = document.createElement("span");
 
-let burgers;
 
-const buscarProductosEnJSON = (burgersABuscar) => {
-    fetch("productos.json")
-    .then(response => response.json())
-    .then(data => {
-        burgers = data;
-        console.log(burgers);
-        //generadorCards(burgers)
-    })
-};
 
-for (let i = 0; i < productos.length; i++) {
-    if (productos[i].stock) {
-        acumulador += `<div class="container px-4 px-lg-5 mt-5">
+async function obtenerProductos(){
+    const productos = await fetch("./productos.json")
+    .then(res => res.json())
+    .then(data =>  data.burgers);
+    return productos;
+}
+
+async function listaDeProductos(){
+    const productos =  await obtenerProductos();
+    mostrarProdcutos(productos);
+}
+
+function mostrarProdcutos(productos) {
+    const listado = document.getElementById("cardProduct");
+    productos.forEach((producto )=>{
+        const {name, description,id, price, image} = producto;
+        const acumulador = `<div class="container px-4 px-lg-5 mt-5" style="width: 30%">
         <div class="row mb-5">
          <div class="card h-100">
             <!-- Sale badge-->
             <div class="badge bg-dark text-white position-absolute" style="top: 0.5rem; right: 0.5rem">Sale</div>
             <!-- Product image-->
-            <img class="card-img-top" src="https://dummyimage.com/450x300/dee2e6/6c757d.jpg" alt="..." />
+            <img class="card-img-top" src="${image}" alt="..." />
             <!-- Product details-->
             <div class="card-body p-4">
                 <div class="text-center">
                     <!-- Product name-->
-                    <h5 class="fw-bolder">${productos[i].burger}</h5>
+                    <h5 class="fw-bolder">${name}</h5>
                     <!-- Product reviews-->
                     <div class="d-flex justify-content-center small text-warning mb-2">
                         <div class="bi-star-fill"></div>
@@ -47,23 +45,31 @@ for (let i = 0; i < productos.length; i++) {
                         <div class="bi-star-fill"></div>
                         <div class="bi-star-fill"></div>
                     </div>
+                    <h6 class="fw-bolder">${description}</h6>
                     <!-- Product price-->
-                    <span class="text-muted text-decoration-line-through">$20.00</span>
-                    $${productos[i].price}
+                    <span class="text-muted ">$${price}</span>
                 </div>
             </div>
             <!-- Product actions-->
             <div  class="card-footer p-4 pt-0 border-top-0 bg-transparent text-center">
-                <div id="btn-add" onclick="agregarAlCarrito(${productos[i].id})"><a class="btn btn-outline-dark mt-auto" >Agregar al carrito</a></div>
+                <div id="btn-add" onclick="agregarAlCarrito(${id})"><a class="btn btn-outline-dark mt-auto" >Agregar al carrito</a></div>
             </div>
             <div  class="card-footer p-4 pt-0 border-top-0 bg-transparent text-center">
-                <div id="btn-delete" onclick="eliminarDelCarrito(${productos[i].id})"><a class="btn btn-outline-dark mt-auto" >Eliminar del carrito</a></div>
+                <div id="btn-delete" onclick="eliminarDelCarrito(${id})"><a class="btn btn-outline-dark mt-auto" >Eliminar del carrito</a></div>
             </div>
         </div>
         </div>
         </div>`;
-    }
+        listado.innerHTML += acumulador;
+    });
 }
+
+listaDeProductos().then();
+
+
+
+
+
 
 // Reduce para saber el precio total en el carrito
 function precioTotal(productosEnElCarrito) {
@@ -71,13 +77,14 @@ function precioTotal(productosEnElCarrito) {
     
 }
 
-const cartStorage=localStorage.getItem('carrito');
+const cartStorage=localStorage.getItem('carrito').split(",");
+console.log(cartStorage);
 
 const carrito = JSON.parse(cartStorage) ??  [];
 
 
-function agregarAlCarrito(idDeProducto){
-    
+async function agregarAlCarrito(idDeProducto){
+    const productos =  await obtenerProductos();
     //Comparo ID y push
     const indiceEncontrado = productos.findIndex(producto => producto.id == idDeProducto);
     productosEnElCarrito.push(productos[indiceEncontrado]);
@@ -93,13 +100,13 @@ function agregarAlCarrito(idDeProducto){
     Swal.fire({
         position: 'top-end',
         icon: 'success',
-        title: 'Your work has been saved',
+        title: 'Has agregado una hamburguesa al carrito',
         showConfirmButton: false,
         timer: 1500
     })
 
 
-    document.getElementById("cart").innerHTML =  `<span> Carrito:${listadoCarrito.length} - $${total}</span>`;
+    document.getElementById("cart").innerHTML =  `<span> Carrito:${cartStorage.length} - $${total}</span>`;
     
     //alert("Usted tiene "+ listadoCarrito.length + " producto/s en el carrito actual: " + listadoCarrito);
     
@@ -112,35 +119,38 @@ function agregarAlCarrito(idDeProducto){
 * 2) Siempre se elimina el ultimo producto agregado pero si el cliente quiere agregar muchos productos y 
 * borrar uno en especifico no puede. 
 */
-function eliminarDelCarrito(idDeProducto){
-    
+async function eliminarDelCarrito(idDeProducto){
+    const productos =  await obtenerProductos();
     //Comparo ID y pop
     const indiceEncontrado = productos.findIndex(producto => producto.id == idDeProducto);
     //Sweetalert2 del eliminarDelCarrito
-    Swal.fire({
-        title: 'Estás seguro?',
-        text: "No podrás volver atrás",
-        icon: 'warning',
-        showCancelButton: true,
-        cancelButtonText: "Cancelar",
-        confirmButtonColor: '#3085d6',
-        cancelButtonColor: '#d33',
-        confirmButtonText: 'Si, borralo!'
-      }).then((result) => {
-        if (result.isConfirmed) {
-            Swal.fire(
-                'Deleted!',
-                'Your file has been deleted.',
-                'success'
-                )
-            }
-            productosEnElCarrito.pop(productos[indiceEncontrado]);
-            // Paso de obj a txt para mandarlo al localStorage
-            const carritoJSON = JSON.stringify(productosEnElCarrito);
-            localStorage.setItem("carrito", carritoJSON);
-            let total = precioTotal(productosEnElCarrito);
-            document.getElementById("cart").innerHTML =  `<span> Carrito:${listadoCarrito.length} - $${total}</span>`;
-      })
+    if(carrito != []){
+        Swal.fire({
+            title: 'Estás seguro?',
+            text: "No podrás volver atrás",
+            icon: 'warning',
+            showCancelButton: true,
+            cancelButtonText: "Cancelar",
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Si, borralo!'
+          }).then((result) => {
+            if (result.isConfirmed) {
+                Swal.fire(
+                    'Borrada!',
+                    'La hamburguesa a sido borrada correctamente!',
+                    'success'
+                    )
+                }
+                productosEnElCarrito.pop(productos[indiceEncontrado]);
+                // Paso de obj a txt para mandarlo al localStorage
+                const carritoJSON = JSON.stringify(productosEnElCarrito);
+                localStorage.setItem("carrito", carritoJSON);
+                let total = precioTotal(productosEnElCarrito);
+                document.getElementById("cart").innerHTML =  `<span> Carrito:${cartStorage.length} - $${total}</span>`;
+          })
+
+    }
 
     listadoCarrito.pop(productos[indiceEncontrado].burger);
     
@@ -153,10 +163,10 @@ function eliminarDelCarrito(idDeProducto){
         
 }
 
-buscarProductosEnJSON();
+
 
 document.body.append(padreCarrito);
-document.getElementById("cardProduct").innerHTML = acumulador;
+//document.getElementById("cardProduct").innerHTML = acumulador;
 
 // const btn = document.getElementById("#btn-add");
 // let btnAgregar = btn.addEventListener("click",()=>{
